Refactor some APIs and address review comments
