+++ coral-visualization/build.gradle
@@ -1,13 +1,14 @@
 dependencies {
-  compile project(path: ':coral-hive')
+  compile project(path: ':coral-common')
   compile (deps.'graph-vis')
-  compile (deps.'hive'.'hive-exec-core') {
+  testCompile project(path: ':coral-hive')
+  testCompile (deps.'hive'.'hive-exec-core') {
     exclude group: 'org.apache.avro', module: 'avro-tools'
-    // These exclusions are required to prevent duplicate classes since we include
-    // shaded jar above
+    // This exclusion is required to prevent duplicate classes since we include
+    // shaded 'calcite-core' transitively
     exclude group: 'org.apache.calcite', module: 'calcite-core'
     exclude group: 'org.apache.calcite', module: 'calcite-avatica'
   }
-  compile deps.'hadoop'.'hadoop-mapreduce-client-core'
-  compile deps.'kryo'
+  testCompile deps.'hadoop'.'hadoop-mapreduce-client-core'
+  testCompile deps.'kryo'
 }
